type gate arguments as int rather than float (#1315)

+++ pyquil/external/rpcq.py
@@ -18,7 +18,7 @@ class MeasureInfo(Operator):
 
 class GateInfo(Operator):
     parameters: List[Union[float, str]] = Field(default_factory=list)
-    arguments: List[Union[float, str]] = Field(default_factory=list)
+    arguments: List[Union[int, str]] = Field(default_factory=list)
     operator_type: Literal["gate"] = "gate"
